pyproject.toml
[tool.poetry]
name = "funcnodes-umap"
version = "0.1.0"
description = ""
authors = ["Julian Kimmig <julian.kimmig@gmx.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
umap-learn = "^0.5.6"
funcnodes = "^0.2.14"
funcnodes-plotly = "^0.1.3"

[tool.poetry.group.dev.dependencies]
funcnodes-pandas = ">=0.1.7"
pytest = "^8.2.0"
matplotlib = "^3.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
